feat(app): extend Instant Nav pilot with metadata and privacy/daily
- Add App root viewport/icons/Material Icons parity with _document
- Mark internal-marketing routes noindex with canonical URLs to public paths
- Extract PrivacyPageView and DailyPageView; add App pilot routes
- Force-dynamic pilot layout to avoid SSG timeouts on daily data hooks
- Remove unused @trpc/next; update Instant Nav TODO and AGENTS.md
