+++ package.json
@@ -71,7 +71,6 @@
     "@reduxjs/toolkit": "^2.11.2",
     "@tanstack/react-query": "^5.101.4",
     "@trpc/client": "11.18.0",
-    "@trpc/next": "11.18.0",
     "@trpc/react-query": "11.18.0",
     "@trpc/server": "11.18.0",
     "@upstash/redis": "^1.38.1",
