+++ vibe-docs/Instant-Navigations-TODO.md
@@ -22,11 +22,13 @@
 - [x] Dual-router shell (`next/compat/router`) so App pilot does not throw
 - [x] Public `/` / `/{locale}` stay on Pages `pages/index` (Pages `i18n` cannot rewrite bare locales into App routes)
 - [x] `proxy.ts`: `/api/config` + locale header for direct `/internal-marketing/*`
+- [x] App layout metadata (viewport, icons, Material Icons)
+- [x] `robots: noindex` on pilot routes (`internalMarketingPilotMetadata`)
+- [x] Extend pilot to `/internal-marketing/[locale]/privacy` and `/daily`
 - [ ] Public cutover of home to App (blocked on removing `next.config` `i18n`)
-- [ ] App layout metadata (viewport, icons, Material Icons)
 - [ ] `cacheComponents` / `partialPrefetching` (blocked: root `headers()` + need 16.3)
-- [ ] `unstable_instant` on pilot routes
-- [ ] Extend to `/privacy`, `/daily`
+- [ ] `unstable_instant` on pilot routes (blocked until `cacheComponents`)
+- [x] Remove unused `@trpc/next` dependency
 
 ## Phase 3+ — Playground / full migration
